improve package layout
