+++ backend/database/database.go
@@ -1,7 +1,5 @@
 package database
 
-// TODO: find the best place/package to store this file (think about design)
-
 import (
 	"database/sql"
 
@@ -10,16 +8,13 @@ import (
 	log "github.com/Sirupsen/logrus"
 )
 
-type DatabaseAccessor interface {
-	UserDataAccessor
-	TweetDataAccessor
-}
-
+// Struct that implements DatabaseAccessor
 type Database struct {
 	UserDataAccessor
 	TweetDataAccessor
 }
 
+// Constructs new Database that uses given sql.DB connection
 func NewDatabase(databaseConnection *sql.DB) DatabaseAccessor {
 	return &Database{
 		NewUserDB(databaseConnection),
@@ -27,8 +22,9 @@ func NewDatabase(databaseConnection *sql.DB) DatabaseAccessor {
 	}
 }
 
-func NewDatabaseConnection() *sql.DB {
-	// TODO: read user data, host and port from config file or something like that
+// Returns new connection to DB specified in config file. Panics when unrecoverable error occurs.
+func NewConnection() *sql.DB {
+	// TODO: read user data, host and port from config file
 
 	db, err := sql.Open("postgres", "user=postgres password=postgres host=localhost sslmode=disable")
 	if err != nil {
